+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix typo in ecsService _wait_on_task
 - Fix wait option for ecs resources finishing immediately
+- Handle pods without a clear duplo service controlling them gracefully
 
 ## [0.3.8] - 2025-11-07